Take a different approach to error handling
The 'error' token is tempting but not the right solution (since it
doesn't nudge the buffer forward, we might end up in an infinite loop!).

Instead, I've gone the route that identifiers are basically any word
that _isn't_ otherwise an operator. That does mean some things don't
lex strictly as Haskell, but the divergence should accomodate more
"custom" `Show` instances. Example of things that are now considered
to be just 1 identifier:

  * `foo-bar`
  * `<function>`
  * `tmp-1234`